static language section

+++ app/Http/Controllers/Admin/AdminAuthorController.php
@@ -112,9 +112,17 @@ class AdminAuthorController extends Controller
                 'photo' => 'image|mimes:jpg,jpeg,png,gif'
             ]);
 
-            if ($author_data->photo != NULL) {
+            if (file_exists(public_path('uploads/' . $author_data->photo)) && $author_data->photo != NULL) {
                 unlink(public_path('uploads/' . $author_data->photo));
             }
+
+            // if(file_exists(public_path('uploads/'. $author_data->photo))){
+            //     unlink(public_path('uploads/' . $author_data->photo));
+            // }
+
+            // if ($author_data->photo != NULL) {
+            //     unlink(public_path('uploads/' . $author_data->photo));
+            // }
 
 
             $now = time();
